Develop (#3)
* Add project config

* add 2 exercises

* test mypy

* Finish exercises

* add last exercise

* Update README.md

* add space to readme

* Update errors found by Emiel

+++ src/data_model.py
@@ -5,7 +5,7 @@ def increment(x: int) -> int:
 
 
 class Employee:
-    def __init__(self, name: str, age: int, salary: float) -> None:
+    def __init__(self, name: str, age: int, salary: int) -> None:
         self.name = name
         self.age = age
         self.salary = salary
